feat: 2.2-A estructura — 3er sub-tab "Datos básicos" + reorder + dedup
Etapa 2.2-A (estructural, behavior-preserving). "Mi taller" pasa de 2 a 3
sub-tabs en orden secuencial obligatorio → productivo → curatorial:

  Datos básicos (índice) → Mi gestión productiva → Mi vidriera

- Nuevo sub-tab "Datos básicos" = ruta índice /taller/perfil. Contiene la
  identidad pública (descripción, año, ubicación — se LEEN en la vidriera,
  misma fila Taller, sin copia ni sync: spec §3), los Datos del responsable
  (PII privada, movida desde gestión) y los Datos de registro (CUIT, pedidos).
- "Mi vidriera" se mueve del índice a /taller/perfil/vidriera (contenido intacto).
- "Mi gestión productiva" queda sólo con el Perfil productivo: se quitan las
  cards "Datos del responsable" e "Información General" (dedup, spec §4).
- perfil-header-tabs: 3 entradas en el nuevo orden + mostrarCromo para /vidriera.
  La lógica de tab activa (índice ===, resto startsWith) no genera doble-activo.

Sin lógica de visibilidad (toggles, flag modeloB_revisado, render condicional):
eso es 2.2-B/C. Edición de datos básicos vía el form existente /editar.
